refactor: rename make-dataset-from-array.ts to make-dataset.ts
